refactor: Moved CanvasToolbar to separate component, added horizontal divider
